backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
PyYAML>=6.0,<7.0
pydantic>=2.8,<3.0
requests>=2.31,<3.0
playwright>=1.45,<2.0
ollama>=0.6,<1.0
python-multipart>=0.0.9,<1.0
